Generator Fixes, Jutland Split
A broken Conditional in Generator was throwing errors in all of the imports.
Also fixes the Sturmcoil Error in Imperium.
Complete overhaul of the Patron system for ease of use and improved constraints
Split Jutland off into its own catalog since it has so many units and none of them are usable in normal Imperium Battlefleets.